+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 3.0.0 - 2024-05-05
+
+### Added
+- 4 new Stages (canyon, mountain, ocean, terrace)
+- 1 new character (Tingle)
+- Select stage (or random) in setup
+- Water tiles and death by water
+
+### Changed
+- Greatly revamped CPU controls with pathfinding and smarter state-switching
+
 ## 2.0.0 - 2024-05-03
 
 ### Added